move all user related methods (incl password and tickets) from RODSConn to RC_user trait
